First attempt at querying invalidation, and returning its status
This change queries invalidations that exist in the database, and
return their status during page load for the admin.

If the status is done, they are removed from the database and
returned in the page load so they show up once.

Multiple invals can occur as you drag & drop files. (Each drag &
drop event triggers a new invalidation)